+++ oneservice.properties
@@ -15,3 +15,5 @@ master.datasource.driverClassName=
 master.datasource.url=
 master.datasource.username=
 master.datasource.passwd=
+
+eureka.client.serviceUrl.defaultZone=${EUREKA_URI:http://localhost:8761/eureka}
