+++ pom.xml
@@ -1,49 +1,35 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>com.kj.oneservice.$packageName$</groupId>
-	<artifactId>$servicePath$</artifactId>
-	<version>0.0.1</version>
-	<packaging>jar</packaging>
-
-	<name>$servicePathcc$</name>
-	<description>OneService $servicePath$ Microservice</description>
-
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.5.10.RELEASE</version>
-		<relativePath />
-	</parent>
-
-	<properties>
+<?xml version="1.0" encoding="UTF-8"?><project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.springframework.boot</groupId>
+    <artifactId>spring-boot-starter-parent</artifactId>
+    <version>2.1.6.RELEASE</version>
+  </parent>
+  <groupId>com.pss.oneservice.$packageName$</groupId>
+  <artifactId>$servicepath$</artifactId>
+  <version>0.0.1-SNAPSHOT</version>
+  
+  <properties>
 		<java.version>1.8</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
-
+	
 	<dependencies>
-		<dependency>
-			<groupId>com.kj.oneservice.common</groupId>
+        <dependency>
+        	<groupId>com.pss.oneservice.common</groupId>
 			<artifactId>oneservice-common</artifactId>
-			<version>0.0.1</version>
-		</dependency>
-
-		<dependency>
+			<version>1.0.1</version>
+        </dependency>
+        
+        <dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-jdbc</artifactId>
-			<version>4.0.3.RELEASE</version>
-		</dependency>
-	</dependencies>
-
-	<build>
+		
+    </dependencies>
+    
+    <build>
 		<defaultGoal>clean package</defaultGoal>
 		<plugins>
 			<plugin>
@@ -57,7 +43,7 @@
 			<plugin>
 				<groupId>org.jacoco</groupId>
 				<artifactId>jacoco-maven-plugin</artifactId>
-				<version>0.8.0</version>
+				<version>0.8.4</version>
 				<executions>
 					<execution>
 						<goals>
@@ -79,7 +65,7 @@
 				<version>1.0.1</version>
 				<configuration>
 					<sourceDirectory>${basedir}/src/main/resources/pojoschema</sourceDirectory>
-					<targetPackage>com.kj.oneservice.$packageName$.integration.model</targetPackage>
+					<targetPackage>com.pss.oneservice.$packageName$.integration.model</targetPackage>
 					<outputDirectory>${basedir}/src/main/java</outputDirectory>
 				</configuration>
 				<executions>
@@ -92,20 +78,29 @@
 			</plugin>
 		</plugins>
 	</build>
-	<repositories>
-		<repository>
-			<id>spring-releases</id>
-			<url>https://repo.spring.io/libs-release</url>
-		</repository>
-		<repository>
-			<id>repository.springframework.maven.milestone</id>
-			<url>http://repo.spring.io/milestone</url>
-		</repository>
-	</repositories>
-	<pluginRepositories>
-		<pluginRepository>
-			<id>spring-releases</id>
-			<url>https://repo.spring.io/libs-release</url>
-		</pluginRepository>
-	</pluginRepositories>
+
+    <!-- Add Spring repositories -->
+    <!-- (you don't need this if you are using a .RELEASE version) -->
+    <repositories>
+        <repository>
+            <id>spring-snapshots</id>
+            <url>https://repo.spring.io/snapshot</url>
+            <snapshots><enabled>true</enabled></snapshots>
+        </repository>
+        <repository>
+            <id>spring-milestones</id>
+            <url>https://repo.spring.io/milestone</url>
+        </repository>
+    </repositories>
+    <pluginRepositories>
+        <pluginRepository>
+            <id>spring-snapshots</id>
+            <url>https://repo.spring.io/snapshot</url>
+        </pluginRepository>
+        <pluginRepository>
+            <id>spring-milestones</id>
+            <url>https://repo.spring.io/milestone</url>
+        </pluginRepository>
+    </pluginRepositories>
+    
 </project>
