--- src/main/java/com/kj/oneservice/$packageName$/integration/dao/$serviceNamecc$DAO.java
@@ -1,33 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.dao;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.stereotype.Repository;
-
-import com.kj.oneservice.$packageName$.integration.model.$serviceNamecc$Request;
-
-/**
- * DAO Class to access DB for $serviceNamecc$ service
- * 
- * @author $author$
- */
-@Repository
-public class $serviceNamecc$DAO extends OneServiceDAO {
-
-	/**
-	 * <!--DAO method stub to write database calls. This method can be
-	 * replicated for making more DB calls.-->
-	 * 
-	 * @param $serviceName$Request
-	 * @param queryString
-	 * @return List<Object> objectList
-	 */
-	public List<Object> get($serviceNamecc$Request $serviceName$Request, String queryString) {
-		List<Object> objectList = new ArrayList<Object>();
-
-		// TODO Write the DAO and row mapping code here.
-
-		return objectList;
-	}
-}
