--- src/main/java/com/kj/oneservice/$packageName$/integration/model/$serviceNamecc$Request.java
@@ -1,13 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.model;
-
-import com.kj.oneservice.common.integration.model.ServiceRequest;
-
-/**
- * Request Bean class for $serviceNamecc$ service
- * 
- * @author $author$
- */
-public class $serviceNamecc$Request extends ServiceRequest {
-
-	// TODO Add request variables here
-}
