--- src/main/java/com/kj/oneservice/$packageName$/integration/model/$serviceNamecc$Response.java
@@ -1,13 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.model;
-
-import com.kj.oneservice.common.integration.model.ServiceResponse;
-
-/**
- * Response Bean class for $serviceNamecc$ service
- * 
- * @author $author$
- */
-public class $serviceNamecc$Response extends ServiceResponse {
-
-	// TODO Add response variables here
-}
