--- src/main/java/com/kj/oneservice/$packageName$/integration/service/OneServiceService.java
@@ -1,12 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.service;
-
-/**
- * Service abstract class for $servicePathcc$ service
- * 
- * <!-- This Class DOES NOT require any modification.-->
- * 
- * @author KiranB
- */
-public abstract class OneServiceService {
-
-}
