--- src/main/java/com/kj/oneservice/$packageName$/integration/util/$servicePathcc$Utility.java
@@ -1,10 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.util;
-
-/**
- * Utility class for $servicePathcc$ service
- * 
- * @author $author$
- */
-public class $servicePathcc$Utility {
-
-}
