--- src/main/java/com/kj/oneservice/$packageName$/integration/util/SwaggerConstants.java
@@ -1,11 +1,0 @@
-package com.kj.oneservice.$packageName$.integration.util;
-
-/**
- * Class to hold all Swagger related constants
- * 
- * @author $author$
- */
-public class SwaggerConstants {
-
-	// TODO Add Swagger constants here
-}
