+++ src/main/java/com/pss/oneservice/$packageName$/integration/model/$serviceNamecc$Response.java
@@ -1,0 +1,89 @@
+
+package com.pss.oneservice.$packageName$.integration.model;
+
+import java.util.HashMap;
+import java.util.Map;
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.pss.oneservice.common.integration.model.ServiceResponse;
+
+
+/**
+ * 
+ * 
+ */
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+
+})
+public class $serviceNamecc$Response
+    extends ServiceResponse
+{
+
+    @JsonIgnore
+    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+    @JsonAnyGetter
+    public Map<String, Object> getAdditionalProperties() {
+        return this.additionalProperties;
+    }
+
+    @JsonAnySetter
+    public void setAdditionalProperty(String name, Object value) {
+        this.additionalProperties.put(name, value);
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        sb.append($serviceNamecc$Response.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        int baseLength = sb.length();
+        String superString = super.toString();
+        if (superString!= null) {
+            int contentStart = superString.indexOf('[');
+            int contentEnd = superString.lastIndexOf(']');
+            if ((contentStart >= 0)&&(contentEnd >contentStart)) {
+                sb.append(superString, (contentStart + 1), contentEnd);
+            } else {
+                sb.append(superString);
+            }
+        }
+        if (sb.length()>baseLength) {
+            sb.append(',');
+        }
+        sb.append("additionalProperties");
+        sb.append('=');
+        sb.append(((this.additionalProperties == null)?"<null>":this.additionalProperties));
+        sb.append(',');
+        if (sb.charAt((sb.length()- 1)) == ',') {
+            sb.setCharAt((sb.length()- 1), ']');
+        } else {
+            sb.append(']');
+        }
+        return sb.toString();
+    }
+
+    @Override
+    public int hashCode() {
+        int result = 1;
+        result = ((result* 31)+((this.additionalProperties == null)? 0 :this.additionalProperties.hashCode()));
+        result = ((result* 31)+ super.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+        if ((other instanceof $serviceNamecc$Response) == false) {
+            return false;
+        }
+        $serviceNamecc$Response rhs = (($serviceNamecc$Response) other);
+        return (super.equals(rhs)&&((this.additionalProperties == rhs.additionalProperties)||((this.additionalProperties!= null)&&this.additionalProperties.equals(rhs.additionalProperties))));
+    }
+
+}
