+++ src/main/java/com/pss/oneservice/$packageName$/integration/service/OneServiceService.java
@@ -1,0 +1,12 @@
+package com.pss.oneservice.$packageName$.integration.service;
+
+/**
+ * Service abstract class for $servicepathcc$ service
+ * 
+ * <!-- This Class DOES NOT require any modification.-->
+ * 
+ * @author KiranB
+ */
+public abstract class OneServiceService {
+
+}
