+++ src/main/java/com/pss/oneservice/$packageName$/integration/util/$servicePathcc$Utility.java
@@ -1,0 +1,10 @@
+package com.pss.oneservice.$packageName$.integration.util;
+
+/**
+ * Utility class for $servicePathcc$ service
+ * 
+ * @author $author$
+ */
+public class $servicePathcc$Utility {
+
+}
