+++ src/main/java/com/pss/oneservice/$packageName$/integration/util/SwaggerConstants.java
@@ -1,0 +1,11 @@
+package com.pss.oneservice.$packageName$.integration.util;
+
+/**
+ * Class to hold all Swagger related constants
+ * 
+ * @author $author$
+ */
+public class SwaggerConstants {
+
+	// TODO Add Swagger constants here
+}
