--- src/main/resources/application.properties
@@ -1,50 +1,0 @@
-endpoints.beans.id=springbeans
-endpoints.beans.sensitive=false
-endpoints.beans.enabled=true
-
-endpoints.enabled=false
-endpoints.health.enabled=true
-endpoints.metrics.enabled=true
-endpoints.info.enabled=true
-endpoints.shutdown.enabled=true
-
-endpoints.health.sensitive=false
-endpoints.metrics.sensitive=false
-endpoints.shutdown.sensitive=true
-
-server.port=${server.port}
-#security.user.name=${app.who}
-#security.user.password=${app.me}
-#management.port=${admin.port}
-#management.security.role=SUPERUSER
-management.security.enabled=true
-management.security.roles=ADMIN
-info.app.name=${app.infoname}
-info.app.description=${app.infodesc}
-info.app.version=${app.infover}
-#spring.datasource.max-active=100 # Advanced configuration...
-#spring.datasource.max-idle=8
-#spring.datasource.min-idle=8
-#spring.datasource.initial-size=10
-#spring.datasource.validation-query=
-#spring.datasource.test-on-borrow=false
-#spring.datasource.test-on-return=false
-#spring.datasource.test-while-idle=
-#spring.datasource.time-between-eviction-runs-millis=
-#spring.datasource.min-evictable-idle-time-millis=
-#spring.datasource.max-wait=
-# BANNER
-banner.charset=UTF-8
-# Banner file encoding.
-banner.location=file:${app.home}/banner/${server.port}-banner.txt
-# Banner file location.
-banner.image.location=file:${app.home}/banner/${server.port}-banner.gif
- # Banner image file location (jpg/png can also be used).
-banner.image.width=76
-# Width of the banner image in chars (default 76)
-banner.image.height=
-# Height of the banner image in chars (default based on image height)
-banner.image.margin=2
-# Left hand image margin in chars (default 2)
-banner.image.invert=false
-# If images should be inverted for dark terminal themes (default false)
