--- src/main/resources/log4j.properties
@@ -1,20 +1,0 @@
-# Root logger option
-log4j.rootLogger=${app.logger}, stdout,file
-
-log4j.logger.com.kj=${app.logger}, stdout,file
-log4j.additivity.com.kj=false
-
-# Redirect log messages to console
- log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-# log4j.appender.stdout.Target=System.out
-log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern={"RequestId":"%X{RequestId}","TimeStamp":"%d{dd-MMM-yyyy HH:mm:ss}","MsgLevel":"%-5p","ClassName":"%c{1}:%L","Message":"%m"}%n
-
-# Redirect log messages to a log file
-log4j.appender.file=org.apache.log4j.DailyRollingFileAppender
-log4j.appender.file.File=${app.logpath}/${app.log}
-log4j.appender.file.datePattern='.'yyyy-MM-dd
-log4j.appender.file.append=true
-log4j.appender.file.layout=org.apache.log4j.PatternLayout
-# log4j.appender.file.layout.ConversionPattern=[%d{dd-MMM-yyyy HH:mm:ss}] [%-5p] [%c{1}:%L] %m%n
-log4j.appender.file.layout.ConversionPattern={"RequestId":"%X{RequestId}","TimeStamp":"%d{dd-MMM-yyyy HH:mm:ss}","MsgLevel":"%-5p","ClassName":"%c{1}:%L","Message":"%m"}%n
