+++ src/main/resources/pojoschema/$serviceNamecc$Request.json
@@ -2,12 +2,26 @@
   "$schema": "http://json-schema.org/draft-04/schema#",
   "description": "",
   "type": "object",
-  "extendsJavaClass" : "com.kj.oneservice.common.integration.model.ServiceRequest",
-  "javaJsonView": "com.kj.oneservice.common.integration.model.ServiceRequest",
-  "minProperties": 0,
-  "maxProperties": 0,
+  "extendsJavaClass" : "com.pss.oneservice.common.integration.model.ServiceRequest",
+  "javaJsonView": "com.pss.oneservice.common.integration.model.ServiceRequest",
+  "minProperties": 2,
+  "maxProperties": 2,
   "properties": {
+    "JUR": {
+      "id": "/properties/JUR",
+      "type": "string",
+      "minLength": 2,
+      "maxLength": 2
+    },
+    "WC": {
+      "id": "/properties/WC",
+      "type": "string",
+      "minLength": 4,
+      "maxLength": 4
+    }
   },
   "required": [
+    "JUR",
+    "WC"
   ]
 }
