+++ src/main/resources/pojoschema/$serviceNamecc$Response.json
@@ -2,8 +2,8 @@
   "$schema": "http://json-schema.org/draft-04/schema#",
   "description": "",
   "type": "object",
-  "extendsJavaClass" : "com.kj.oneservice.common.integration.model.ServiceResponse",
-  "javaJsonView": "com.kj.oneservice.common.integration.model.ServiceResponse",
+  "extendsJavaClass" : "com.pss.oneservice.common.integration.model.ServiceResponse",
+  "javaJsonView": "com.pss.oneservice.common.integration.model.ServiceResponse",
   "minProperties": 0,
   "maxProperties": 0,
   "properties": {
